+++ FSI/FSI.C
@@ -55,6 +55,21 @@ bool preciceAdapter::FSI::FluidStructureInteraction::readConfig(const IOdictiona
     solverType_ = FSIdict.lookupOrDefault<word>("solverType", "");
     DEBUG(adapterInfo("    user-defined solver type : " + solverType_));
 
+    // When restarting FSI simulations, we may need to account for previous displacement.
+    // We do this by resetting the displacement when defining the interface.
+    // Since this is a feature that may not work as expected, depending on the implementation of the
+    // structure solver used, we make this feature opt-in.
+    restartFromDeformed_ = FSIdict.lookupOrDefault<bool>("restartFromDeformed", true);
+
+    if (solverType_ == "solid" && !restartFromDeformed_)
+    {
+        adapterInfo("The option \"restartFromDeformed\" is only valid for Fluid solvers. Solid solvers usually use exclusively the reference configuration for computations and the mesh is not deforming over time.", "error");
+    }
+
+    if (solverType_ != "solid")
+    {
+        DEBUG(adapterInfo("    restart from deformed : " + std::to_string(restartFromDeformed_)));
+    }
     /* TODO: Read the names of any needed fields and parameters.
      * Include the force here?
      */
@@ -67,9 +82,9 @@ bool preciceAdapter::FSI::FluidStructureInteraction::readConfig(const IOdictiona
     nameCellDisplacement_ = FSIdict.lookupOrDefault<word>("nameCellDisplacement", "cellDisplacement");
     DEBUG(adapterInfo("    cellDisplacement field name : " + nameCellDisplacement_));
 
-    // Read the name of the solidForce field (if different)
-    nameSolidForce_ = FSIdict.lookupOrDefault<word>("nameSolidForce", "solidForce");
-    DEBUG(adapterInfo("    solidForce field name : " + nameSolidForce_));
+    // Read the name of the force field (if different)
+    nameForce_ = FSIdict.lookupOrDefault<word>("nameForce", "Force");
+    DEBUG(adapterInfo("    force field name : " + nameForce_));
 
     return true;
 }
@@ -123,7 +138,7 @@ bool preciceAdapter::FSI::FluidStructureInteraction::addWriters(std::string data
     {
         interface->addCouplingDataWriter(
             dataName,
-            new Force(mesh_, solverType_, nameSolidForce_) /* TODO: Add any other arguments here */
+            new Force(mesh_, solverType_, nameForce_) /* TODO: Add any other arguments here */
         );
         DEBUG(adapterInfo("Added writer: Force."));
     }
@@ -171,7 +186,7 @@ bool preciceAdapter::FSI::FluidStructureInteraction::addReaders(std::string data
     {
         interface->addCouplingDataReader(
             dataName,
-            new Force(mesh_, solverType_, nameSolidForce_) /* TODO: Add any other arguments here */
+            new Force(mesh_, solverType_, nameForce_) /* TODO: Add any other arguments here */
         );
         DEBUG(adapterInfo("Added reader: Force."));
     }
@@ -209,4 +224,19 @@ bool preciceAdapter::FSI::FluidStructureInteraction::addReaders(std::string data
     // the one provided in the adapter's configuration file.
 
     return found;
+}
+
+std::string preciceAdapter::FSI::FluidStructureInteraction::getCellDisplacementFieldName()
+{
+    return nameCellDisplacement_;
+}
+
+std::string preciceAdapter::FSI::FluidStructureInteraction::getPointDisplacementFieldName()
+{
+    return namePointDisplacement_;
+}
+
+bool preciceAdapter::FSI::FluidStructureInteraction::isRestartingFromDeformed()
+{
+    return restartFromDeformed_;
 }