+++ Interface.C
@@ -11,11 +11,15 @@ preciceAdapter::Interface::Interface(
     std::string meshName,
     std::string locationsType,
     std::vector<std::string> patchNames,
-    bool meshConnectivity)
+    bool meshConnectivity,
+    bool restartFromDeformed,
+    const std::string& namePointDisplacement,
+    const std::string& nameCellDisplacement)
 : precice_(precice),
   meshName_(meshName),
   patchNames_(patchNames),
-  meshConnectivity_(meshConnectivity)
+  meshConnectivity_(meshConnectivity),
+  restartFromDeformed_(restartFromDeformed)
 {
     // Get the meshID from preCICE
     meshID_ = precice_.getMeshID(meshName_);
@@ -72,10 +76,10 @@ preciceAdapter::Interface::Interface(
     }
 
     // Configure the mesh (set the data locations)
-    configureMesh(mesh);
+    configureMesh(mesh, namePointDisplacement, nameCellDisplacement);
 }
 
-void preciceAdapter::Interface::configureMesh(const fvMesh& mesh)
+void preciceAdapter::Interface::configureMesh(const fvMesh& mesh, const std::string& namePointDisplacement, const std::string& nameCellDisplacement)
 {
     // The way we configure the mesh differs between meshes based on face centers
     // and meshes based on face nodes.
@@ -91,6 +95,12 @@ void preciceAdapter::Interface::configureMesh(const fvMesh& mesh)
                 mesh.boundaryMesh()[patchIDs_.at(j)].faceCentres().size();
         }
         DEBUG(adapterInfo("Number of face centres: " + std::to_string(numDataLocations_)));
+
+        // In case we want to perform the reset later on, look-up the corresponding data field name
+        Foam::volVectorField const* cellDisplacement = nullptr;
+        if (mesh.foundObject<volVectorField>(nameCellDisplacement))
+            cellDisplacement =
+                &mesh.lookupObject<volVectorField>(nameCellDisplacement);
 
         // Array of the mesh vertices.
         // One mesh is used for all the patches and each vertex has 3D coordinates.
@@ -108,8 +118,15 @@ void preciceAdapter::Interface::configureMesh(const fvMesh& mesh)
         for (uint j = 0; j < patchIDs_.size(); j++)
         {
             // Get the face centers of the current patch
-            const vectorField faceCenters =
+            vectorField faceCenters =
                 mesh.boundaryMesh()[patchIDs_.at(j)].faceCentres();
+
+            // Move the interface according to the current values of the cellDisplacement field,
+            // to account for any displacements accumulated before restarting the simulation.
+            // This is information that OpenFOAM reads from its result/restart files.
+            // If the simulation is not restarted, the displacement should be zero and this line should have no effect.
+            if (cellDisplacement != nullptr && !restartFromDeformed_)
+                faceCenters -= cellDisplacement->boundaryField()[patchIDs_.at(j)];
 
             // Assign the (x,y,z) locations to the vertices
             for (int i = 0; i < faceCenters.size(); i++)
@@ -157,11 +174,11 @@ void preciceAdapter::Interface::configureMesh(const fvMesh& mesh)
                     }
                     else
                     {
-                        FatalErrorInFunction
-                            << "It seems like you are using preCICE in 2D and your geometry is not located int the xy-plane. "
-                               "The OpenFOAM adapter implementation supports preCICE 2D cases only with the z-axis as out-of-plane direction."
-                               "Please rotate your geometry so that the geometry is located in the xy-plane."
-                            << exit(FatalError);
+                        adapterInfo("It seems like you are using preCICE in 2D and your geometry is not located int the xy-plane. "
+                                    "The OpenFOAM adapter implementation supports preCICE 2D cases only with the z-axis as out-of-plane direction."
+                                    "Please rotate your geometry so that the geometry is located in the xy-plane."
+                                    "If you are running a 2D axisymmetric case just ignore this.",
+                                    "warning");
                     }
                 }
             }
@@ -179,6 +196,12 @@ void preciceAdapter::Interface::configureMesh(const fvMesh& mesh)
                 mesh.boundaryMesh()[patchIDs_.at(j)].localPoints().size();
         }
         DEBUG(adapterInfo("Number of face nodes: " + std::to_string(numDataLocations_)));
+
+        // In case we want to perform the reset later on, look-up the corresponding data field name
+        Foam::pointVectorField const* pointDisplacement = nullptr;
+        if (mesh.foundObject<pointVectorField>(namePointDisplacement))
+            pointDisplacement =
+                &mesh.lookupObject<pointVectorField>(namePointDisplacement);
 
         // Array of the mesh vertices.
         // One mesh is used for all the patches and each vertex has 3D coordinates.
@@ -200,8 +223,20 @@ void preciceAdapter::Interface::configureMesh(const fvMesh& mesh)
             // TODO: Check if this behaves correctly in parallel.
             // TODO: Check if this behaves correctly with multiple, connected patches.
             // TODO: Maybe this should be a pointVectorField?
-            const pointField faceNodes =
+            pointField faceNodes =
                 mesh.boundaryMesh()[patchIDs_.at(j)].localPoints();
+
+            // Similar to the cell displacement above:
+            // Move the interface according to the current values of the cellDisplacement field,
+            // to account for any displacements accumulated before restarting the simulation.
+            // This is information that OpenFOAM reads from its result/restart files.
+            // If the simulation is not restarted, the displacement should be zero and this line should have no effect.
+            if (pointDisplacement != nullptr && !restartFromDeformed_)
+            {
+                const vectorField& resetField = refCast<const vectorField>(
+                    pointDisplacement->boundaryField()[patchIDs_.at(j)]);
+                faceNodes -= resetField;
+            }
 
             // Assign the (x,y,z) locations to the vertices
             // TODO: Ensure consistent order when writing/reading
@@ -239,7 +274,15 @@ void preciceAdapter::Interface::configureMesh(const fvMesh& mesh)
 
                 // Get the list of faces and coordinates at the interface patch
                 const List<face> faceField = mesh.boundaryMesh()[patchIDs_.at(j)].localFaces();
-                const Field<point> pointCoords = mesh.boundaryMesh()[patchIDs_.at(j)].localPoints();
+                Field<point> pointCoords = mesh.boundaryMesh()[patchIDs_.at(j)].localPoints();
+
+                // Subtract the displacement part in case we have deformation
+                if (pointDisplacement != nullptr && !restartFromDeformed_)
+                {
+                    const vectorField& resetField = refCast<const vectorField>(
+                        pointDisplacement->boundaryField()[patchIDs_.at(j)]);
+                    pointCoords -= resetField;
+                }
 
                 // Array to store coordinates in preCICE format
                 double triCoords[faceField.size() * triaPerQuad * nodesPerTria * componentsPerNode];
@@ -443,38 +486,34 @@ void preciceAdapter::Interface::createBuffer()
 
 void preciceAdapter::Interface::readCouplingData()
 {
-    // Are new data available or is the participant subcycling?
-    if (precice_.isReadDataAvailable())
+    // Make every coupling data reader read
+    for (uint i = 0; i < couplingDataReaders_.size(); i++)
     {
-        // Make every coupling data reader read
-        for (uint i = 0; i < couplingDataReaders_.size(); i++)
+        // Pointer to the current reader
+        preciceAdapter::CouplingDataUser*
+            couplingDataReader = couplingDataReaders_.at(i);
+
+        // Make preCICE read vector or scalar data
+        // and fill the adapter's buffer
+        if (couplingDataReader->hasVectorData())
         {
-            // Pointer to the current reader
-            preciceAdapter::CouplingDataUser*
-                couplingDataReader = couplingDataReaders_.at(i);
-
-            // Make preCICE read vector or scalar data
-            // and fill the adapter's buffer
-            if (couplingDataReader->hasVectorData())
-            {
-                precice_.readBlockVectorData(
-                    couplingDataReader->dataID(),
-                    numDataLocations_,
-                    vertexIDs_,
-                    dataBuffer_);
-            }
-            else
-            {
-                precice_.readBlockScalarData(
-                    couplingDataReader->dataID(),
-                    numDataLocations_,
-                    vertexIDs_,
-                    dataBuffer_);
-            }
-
-            // Read the received data from the buffer
-            couplingDataReader->read(dataBuffer_, dim_);
+            precice_.readBlockVectorData(
+                couplingDataReader->dataID(),
+                numDataLocations_,
+                vertexIDs_,
+                dataBuffer_);
         }
+        else
+        {
+            precice_.readBlockScalarData(
+                couplingDataReader->dataID(),
+                numDataLocations_,
+                vertexIDs_,
+                dataBuffer_);
+        }
+
+        // Read the received data from the buffer
+        couplingDataReader->read(dataBuffer_, dim_);
     }
 }